backend/requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv~=1.0.0
pip~=21.2.4
